+++ tools/test/test262/excludelist.orig.xml
@@ -177,8 +177,6 @@
   <test id="attempt-to-redeclare-var-with-function-declaration"><reason>TODO</reason></test>
   <test id="generator-super-call-body"><reason>TODO</reason></test>
   <test id="generator-super-call-param"><reason>TODO</reason></test>
-  <test id="invalid-hexidecimal-character-escape-sequence-truncated"><reason>TODO</reason></test>
-  <test id="invalid-unicode-escape-sequence-truncated"><reason>TODO</reason></test>
   <test id="class-body-has-direct-super-missing-class-heritage"><reason>TODO</reason></test>
   <test id="class-body-method-contains-direct-super"><reason>TODO</reason></test>
   <test id="class-body-special-method-generator-contains-direct-super"><reason>TODO</reason></test>
@@ -187,7 +185,4 @@
   <test id="class-body-static-method-contains-direct-super"><reason>TODO</reason></test>
   <test id="class-body-static-method-get-contains-direct-super"><reason>TODO</reason></test>
   <test id="class-body-static-method-set-contains-direct-super"><reason>TODO</reason></test>
-  <test id="with-initializer-label-statement"><reason>TODO</reason></test>
-  <test id="with-initialisers-in-statement-positions-label-statement"><reason>TODO</reason></test>
-  <test id="without-initialisers-in-statement-positions-label-statement"><reason>TODO</reason></test>
 </excludeList>
